+++ examples/main.py
@@ -1,10 +1,13 @@
+from datetime import date, datetime
+
+import aioredis
 import uvicorn
 from fastapi import FastAPI
 from starlette.requests import Request
 from starlette.responses import Response
 
 from fastapi_cache import FastAPICache
-from fastapi_cache.backends.inmemory import InMemoryBackend
+from fastapi_cache.backends.redis import RedisBackend
 from fastapi_cache.decorator import cache
 
 app = FastAPI()
@@ -20,7 +23,7 @@ async def get_ret():
 
 
 @app.get("/")
-@cache(namespace="test", expire=2)
+@cache(namespace="test", expire=20)
 async def index(request: Request, response: Response):
     return dict(ret=await get_ret())
 
@@ -30,9 +33,22 @@ async def clear():
     return await FastAPICache.clear(namespace="test")
 
 
+@app.get("/date")
+@cache(namespace="test", expire=20)
+async def get_data(request: Request, response: Response):
+    return date.today()
+
+
+@app.get("/datetime")
+@cache(namespace="test", expire=20)
+async def get_datetime(request: Request, response: Response):
+    return datetime.now()
+
+
 @app.on_event("startup")
 async def startup():
-    FastAPICache.init(InMemoryBackend(), prefix="fastapi-cache")
+    redis = aioredis.from_url(url="redis://localhost")
+    FastAPICache.init(RedisBackend(redis), prefix="fastapi-cache")
 
 
 if __name__ == "__main__":