+++ fastapi_cache/backends/redis.py
@@ -9,21 +9,19 @@ class RedisBackend(Backend):
     def __init__(self, redis: Redis):
         self.redis = redis
 
-    async def get_with_ttl(self, name: str) -> Tuple[int, str]:
-        p = self.redis.pipeline()
-        p.ttl(name)
-        p.get(name)
-        return await p.execute()
+    async def get_with_ttl(self, key: str) -> Tuple[int, str]:
+        async with self.redis.pipeline(transaction=True) as pipe:
+            return await (pipe.ttl(key).get(key).execute())
 
     async def get(self, name) -> str:
         return await self.redis.get(name)
 
-    async def set(self, name: str, value: str, ex: int = None):
-        return await self.redis.set(name, value, ex=ex)
+    async def set(self, key: str, value: str, expire: int = None):
+        return await self.redis.set(key, value, ex=expire)
 
-    async def clear(self, namespace: str = None, name: str = None) -> int:
+    async def clear(self, namespace: str = None, key: str = None) -> int:
         if namespace:
             lua = f"for i, name in ipairs(redis.call('KEYS', '{namespace}:*')) do redis.call('DEL', name); end"
-            return await self.redis.eval(lua)
-        elif name:
-            return await self.redis.delete(name)
+            return await self.redis.eval(lua, numkeys=0)
+        elif key:
+            return await self.redis.delete(key)