+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## 0.1
 
+### 0.1.7
+
+- Fix default json coder for datetime.
+
+### 0.1.6
+
+- Fix redis cache.
+- Encode key builder.
+
+### 0.1.5
+
+- Fix setting expire for redis (#24)
+- Update expire key
+
+### 0.1.4
+
+- Fix default expire for memcached. (#13)
+- Update default key builder. (#12)
+
 ### 0.1.3
 
 - Fix cache key builder.